add desc to predefined roles

+++ pkg/microservice/user/core/service/permission/role.go
@@ -350,6 +350,11 @@ func UpdateRole(ns string, req *CreateRoleReq, log *zap.SugaredLogger) error {
 		return fmt.Errorf("failed to find role: [%s] in namespace [%s], error: %s", req.Namespace, ns, err)
 	}
 
+	if roleInfo.Type == int64(setting.RoleTypeSystem) {
+		tx.Rollback()
+		return fmt.Errorf("can't update system role: %s", req.Name)
+	}
+
 	err = orm.DeleteRoleActionBindingByRole(roleInfo.ID, tx)
 	if err != nil {
 		log.Errorf("failed to delete role-action binding for role: %s, error: %s", roleInfo.Name, err)
@@ -598,19 +603,19 @@ func BatchDeleteRole(roles []*models.NewRole, db *gorm.DB, log *zap.SugaredLogge
 func CreateDefaultRolesForNamespace(namespace string, log *zap.SugaredLogger) error {
 	projectAdminRole := &models.NewRole{
 		Name:        "project-admin",
-		Description: "",
+		Description: "拥有指定项目中任何操作的权限",
 		Type:        int64(setting.RoleTypeSystem),
 		Namespace:   namespace,
 	}
 	readOnlyRole := &models.NewRole{
 		Name:        "read-only",
-		Description: "",
+		Description: "拥有指定项目中所有资源的读权限",
 		Type:        int64(setting.RoleTypeSystem),
 		Namespace:   namespace,
 	}
 	readProjectOnlyRole := &models.NewRole{
 		Name:        "read-project-only",
-		Description: "",
+		Description: "拥有指定项目本身的读权限，无权限查看和操作项目内资源",
 		Type:        int64(setting.RoleTypeSystem),
 		Namespace:   namespace,
 	}
